+++ Runtime/Scripts/AssetMgr.cs
@@ -29,11 +29,15 @@ namespace QuickDemo
             }
         }
 
-        public void LoadAsset<T>(string key, Action<T> callback, bool cache = true) where T : UnityEngine.Object
+        public void LoadAssetAsync<T>(string key, Action<T> callback, bool cache = true) where T : UnityEngine.Object
         {
             AsyncOperationHandle op;
             if (operationDictionary.TryGetValue(key, out op))
             {
+                if (showLog)
+                {
+                    Debug.Log($"[ASSET]LoadAssetAsync from cache successed > {key}");
+                }
                 callback?.Invoke((T)op.Result);
                 return;
             }
@@ -42,19 +46,19 @@ namespace QuickDemo
                 if (opHandle.Status != AsyncOperationStatus.Succeeded)
                 {
                     Debug.LogError($"[ASSET]LoadAssetAsync failed > {key}");
+                    Addressables.Release(opHandle);
                     callback?.Invoke(default(T));
                     return;
                 }
                 
                 if (showLog)
+                {
                     Debug.Log($"[ASSET]LoadAssetAsync successed > {key}");
+                }
+                
                 if (cache)
                 {
                     operationDictionary[key] = opHandle;
-                }
-                else
-                {
-                    Addressables.Release(opHandle);
                 }
                 callback?.Invoke(opHandle.Result);
             }));
@@ -67,9 +71,44 @@ namespace QuickDemo
             callback?.Invoke(opHandle);
         }
 
+        public T LoadAsset<T>(string key, bool cache = true) where T : UnityEngine.Object
+        {
+            AsyncOperationHandle opHandle;
+            if (operationDictionary.TryGetValue(key, out opHandle))
+            {
+                if (showLog)
+                {
+                    Debug.Log($"[ASSET]LoadAsset from cache successed > {key}");
+                }
+                return (T)opHandle.Result;
+            }
+
+            opHandle = Addressables.LoadAssetAsync<T>(key);
+            opHandle.WaitForCompletion();
+        
+            if (opHandle.Status == AsyncOperationStatus.Succeeded)
+            {
+                if (showLog)
+                {
+                    Debug.Log($"[ASSET]LoadAsset successed > {key}");
+                }
+
+                if (cache)
+                {
+                    operationDictionary[key] = opHandle;
+                }
+                return (T)opHandle.Result;
+            }
+            else
+            {
+                Addressables.Release(opHandle);
+                return default(T);
+            }
+        }
+
         public static void InstGameObjectAsync(string key, Action<GameObject> callback)
         {
-            AssetMgr.Inst.LoadAsset<GameObject>(key, (x)=>{
+            AssetMgr.Inst.LoadAssetAsync<GameObject>(key, (x)=>{
                 if (x != null)
                 {
                     var obj = GameObject.Instantiate(x);
@@ -81,6 +120,17 @@ namespace QuickDemo
                     callback?.Invoke(null);
                 }
             });
+        }
+
+        public static GameObject InstGameObject(string key)
+        {
+            var asset = AssetMgr.Inst.LoadAsset<GameObject>(key);
+            if (asset != null)
+            {
+                return GameObject.Instantiate(asset);
+            }
+            Debug.Log("[ASSET]failed inst obj > " + key);
+            return null;
         }
     }
 }