+++ Runtime/Scripts/Utils.cs
@@ -1,5 +1,6 @@
 using System;
 using UnityEngine;
+using System.Text;
 
 namespace QuickDemo
 {
@@ -7,8 +8,35 @@ namespace QuickDemo
     {
         public static Vector3 Vector3ZeroY(Vector3 v3)
         {
-            v3.y = 0;
-            return v3;
+            return Vector3Y(v3, 0);
+        }
+
+        public static Vector3 Vector3MaxY(Vector3 v, float y)
+        {
+            return Vector3Y(v, Mathf.Max(v.y, y));
+        }
+
+        public static Vector3 Vector3MinY(Vector3 v, float y)
+        {
+            return Vector3Y(v, Mathf.Min(v.y, y));
+        }
+
+        public static Vector3 Vector3X(Vector3 v, float x)
+        {
+            v.x = x;
+            return v;
+        }
+
+        public static Vector3 Vector3Y(Vector3 v, float y)
+        {
+            v.y = y;
+            return v;
+        }
+
+        public static Vector3 Vector3Z(Vector3 v, float z)
+        {
+            v.z = z;
+            return v;
         }
 
         public static T GetOrAddComponent<T>(this GameObject go) where T : Component
@@ -65,21 +93,44 @@ namespace QuickDemo
         }
 
         // OnGUI
-        public static string GUIButtonText(string text, int size = 40)
+        public static string GUIText(string text, int size = 40)
         {
             return $"<size={size}>{text}</size>";
         }
 
-        public static GUILayoutOption[] GUIButtonOptions(float height)
+        public static GUILayoutOption[] GUIOptions(float height)
         {
             return new GUILayoutOption[]{
                 GUILayout.Height(height)
             };
         }
 
-        public static GUILayoutOption[] DefaultGUIButtonOptions
+        public static GUILayoutOption[] GUIDefaultButtonOptions
         {
-            get { return GUIButtonOptions(70); }
+            get { return GUIOptions(70); }
+        }
+
+        public static GUILayoutOption[] GUIDefaultLabelOptions
+        {
+            get { return GUIOptions(70); }
+        }
+
+        public static StringBuilder sb = new StringBuilder();
+        public static string GetGameObjectPath(GameObject obj)
+        {
+            if (obj == null)
+                return string.Empty;
+
+            sb.Clear();
+            sb.Insert(0, obj.transform.name);
+            var parent = obj.transform.parent;
+            while (parent != null)
+            {
+                sb.Insert(0, "/");
+                sb.Insert(0, parent.name);
+                parent = parent.parent;
+            }
+            return sb.ToString();
         }
     }
 }