+++ CHANGELOG.md
@@ -1,22 +1,30 @@
 # Changelog
 All notable changes to this package will be documented in this file. The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [1.0.4] - 2020-06-23
+*Compatible with Unity 2020.3*
+### Added
+- AssetMgr sync load asset
+### Changed
+- Utils vector3 help function
+- FSM current state update order
+
 ## [1.0.3] - 2020-06-14
 *Compatible with Unity 2020.3*
 ### Changed
-- update Utils GUI function
+- Utils GUI function
 
 ## [1.0.2] - 2020-06-05
 *Compatible with Unity 2020.3*
 ### Changed
-- update CharacterAnimation
+- Character animator
 
 ## [1.0.1] - 2020-06-05
 *Compatible with Unity 2020.3*
 ### Added
 - HideInPlaying
 ### Changed
-- update Utils
+- Utils random function
 
 ## [1.0.0] - 2020-04-24
 *Compatible with Unity 2020.3*